requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2.16
django-bootstrap5==22.2
django-cors-headers==4.5.0
django-debug-toolbar==3.4.0
django-filter==23.1
django-pagination==1.0.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.2.2
Faker==12.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
flake8-isort==6.1.1
idna==3.10
iniconfig==1.1.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mixer==7.2.2
oauthlib==3.2.2
packaging==21.3
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.5.2
PyJWT==2.1.0
pyparsing==3.0.9
pyTelegramBotAPI==4.26.0
pytest==8.3.5
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
pytest-subtests==0.9.0
pytest-timeout==2.3.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.6
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
soupsieve==2.6
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
webcolors==24.11.1
yapf==0.32.0
